Driver sink parameter is now a named tuple
So a driver prototype now takes a namedtuple as input and ouput. The
namedtuple contains only streams. This means that the object returned
by a driver cannot contain a function to call anymore. Instead all
driver requests have to go through one of the sink streams. This change
make the input and output of a driver more coherent.